requirements.txt
fastapi
pydantic
uvicorn
google-generativeai
dotenv